fix project strctures
