+++ lib/app.dart
@@ -1,41 +1,29 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:in_100_days/root.dart';
+import 'package:in_100_days/style/color.dart';
 
-import 'flavors.dart';
-import 'pages/my_home_page.dart';
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
 
-class App extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: F.title,
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: _flavorBanner(
-        child: MyHomePage(),
-        show: kDebugMode,
+    return ProviderScope(
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primaryColor: AppColor.primary,
+          inputDecorationTheme: const InputDecorationTheme(
+            border: UnderlineInputBorder(
+              borderSide: BorderSide(width: 1, color: AppColor.primary),
+            ),
+          ),
+          textSelectionTheme: const TextSelectionThemeData(
+            cursorColor: AppColor.primary,
+          ),
+        ),
+        home: const Root(),
       ),
     );
   }
-
-  Widget _flavorBanner({
-    required Widget child,
-    bool show = true,
-  }) =>
-      show
-          ? Banner(
-        child: child,
-        location: BannerLocation.topStart,
-        message: F.name,
-        color: Colors.green.withOpacity(0.6),
-        textStyle: TextStyle(
-            fontWeight: FontWeight.w700,
-            fontSize: 12.0,
-            letterSpacing: 1.0),
-        textDirection: TextDirection.ltr,
-      )
-          : Container(
-        child: child,
-      );
 }
