+++ lib/main_prod.dart
@@ -1,8 +1,7 @@
-import 'package:flutter/material.dart';
-import 'app.dart';
+import 'package:in_100_days/entrypoint.dart';
 import 'flavors.dart';
 
 void main() {
   F.appFlavor = Flavor.PROD;
-  runApp(App());
+  entrypoint();
 }
